refactored repository code out
